Pass trace-17-
I add __glibc_unlikely() into the if statement in line 53 and 68,
this macro will tell compiler that the statement is "less" likely
to happen, reduce the probability of running unnecessary code.

I also add __glibc_likely() into the if statement in line 447,
I realize that in the auto test program, the chance in failure on
strdup() is more likeiy to happen.